requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
apispec[marshmallow]==5.2.2
    # via flask-smorest
astroid==2.11.6
    # via
    #   pylint
    #   pylint-flask
black==22.3.0
    # via -r requirements.in
certifi==2022.6.15
    # via
    #   minio
    #   requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   black
    #   flask
    #   pip-tools
colorama==0.4.5
    # via
    #   -r requirements.in
    #   click
    #   pylint
dill==0.3.5.1
    # via pylint
flake8==4.0.1
    # via -r requirements.in
flask==2.1.2
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-dotenv
    #   flask-reverse-proxy-fix
    #   flask-smorest
flask-cors==3.0.10
    # via -r requirements.in
flask-dotenv==0.1.2
    # via -r requirements.in
flask-reverse-proxy-fix==0.2.1
    # via -r requirements.in
flask-smorest==0.38.1
    # via -r requirements.in
future-fstrings==1.2.0
    # via pyecore
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.11.4
    # via flask
isort==5.10.1
    # via pylint
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
lazy-object-proxy==1.7.1
    # via astroid
lxml==4.9.2
    # via pyecore
markupsafe==2.1.1
    # via jinja2
marshmallow==3.16.0
    # via
    #   -r requirements.in
    #   apispec
    #   flask-smorest
    #   marshmallow-dataclass
    #   marshmallow-enum
    #   webargs
marshmallow-dataclass[enum]==8.5.8
    # via -r requirements.in
marshmallow-enum==1.5.1
    # via
    #   -r requirements.in
    #   marshmallow-dataclass
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
minio==7.1.9
    # via -r requirements.in
mypy==0.961
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
    #   typing-inspect
ordered-set==4.1.0
    # via pyecore
packaging==21.3
    # via
    #   marshmallow
    #   webargs
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.2
    # via -r requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   pylint
pycodestyle==2.8.0
    # via flake8
pyecore==0.12.1
    # via pyesdl
pyesdl==22.11.2
    # via -r requirements.in
pyflakes==2.4.0
    # via flake8
pylint==2.14.3
    # via
    #   pylint-flask
    #   pylint-plugin-utils
pylint-flask==0.6
    # via -r requirements.in
pylint-plugin-utils==0.7
    # via pylint-flask
pyparsing==3.0.9
    # via packaging
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.28.0
    # via -r requirements.in
restrictedpython==6.0
    # via pyecore
six==1.16.0
    # via flask-cors
structlog==21.5.0
    # via -r requirements.in
tomli==2.0.1
    # via
    #   black
    #   mypy
    #   pep517
    #   pylint
tomlkit==0.11.0
    # via pylint
typing-extensions==4.2.0
    # via
    #   astroid
    #   black
    #   mypy
    #   pylint
    #   typing-inspect
typing-inspect==0.7.1
    # via marshmallow-dataclass
urllib3==1.26.9
    # via
    #   minio
    #   requests
webargs==8.1.0
    # via flask-smorest
werkzeug==2.1.2
    # via
    #   flask
    #   flask-smorest
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via astroid
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
